chore: re-add app-loader, move isValidUrl to react-helpers
