+++ apps/trading/components/app-loader/index.tsx
@@ -1,2 +1,3 @@
-export * from './web3-provider';
 export * from './app-failure';
+export * from './app-loader';
+export * from './web3-provider';
