+++ libs/react-helpers/src/lib/is-valid-url.ts
@@ -1,0 +1,9 @@
+export const isValidUrl = (url?: string) => {
+  if (!url) return false;
+  try {
+    new URL(url);
+    return true;
+  } catch {
+    return false;
+  }
+};
